+++ examples/wirestate-react-signals/src/Application.tsx
@@ -3,7 +3,7 @@ import "./Application.css";
 import {
   useInjection,
   useEvent,
-  useEventsHandler,
+  useAllEvents,
   useEvents,
 } from "@wirestate/react";
 import { useEffect } from "react";
@@ -27,7 +27,7 @@ export function Application() {
     console.info("[Application] Counter was reset (array of events)");
   });
 
-  useEventsHandler((event) => {
+  useAllEvents((event) => {
     console.info(
       "[Application] Log all events:",
       event.type,